Move eigen_matrix_compare.h from maliput_test_utilities to multilane_test_utilities.
