+++ multilane/test/multilane/multilane_connection_test.cc
@@ -9,10 +9,10 @@
 #include <gtest/gtest.h>
 
 #include "maliput/math/vector.h"
-#include "maliput/test_utilities/eigen_matrix_compare.h"
 #include "multilane/arc_road_curve.h"
 #include "multilane/cubic_polynomial.h"
 #include "multilane/line_road_curve.h"
+#include "multilane_test_utilities/eigen_matrix_compare.h"
 #include "multilane_test_utilities/multilane_types_compare.h"
 
 namespace maliput {
